+++ src/main/java/in/koala/service/KeywordService.java
@@ -17,8 +17,8 @@ public interface KeywordService {
 
     List<Notice> getKeywordNotice(String keywordName, String site);
     List<Notice> getSearchNotice(String keywordName, String site, String word);
-    Boolean deleteNotice(List<Integer> noticeList);
-    Boolean deleteNoticeUndo(List<Integer> noticeList);
+    Boolean deleteNotice(List<Long> noticeList);
+    Boolean deleteNoticeUndo(List<Long> noticeList);
     Boolean noticeRead(String noticeId);
 
     List<String> recommendKeyword();
